fix: return same EM fork for EM from core and driver packages
Previously, there were 2 provides registered for the EM:
- one from core package
- one from each installed driver package (knex/mongo)

This brought issues with the nestjs request scopes, as the repositories depend on the core EM,
while users usually inject from the driver package.

Now the driver specific provider just depends on the core EM and returns it, so both injection
tokens behaves the same.